+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # PASSWALL CHANGELOG
 
 ## Version: [1.1.-] (2020-07-**)
+### Add
+- Email configuration
+- SendEmail function to app package
 ### Changed
 - Github organization name pass-wall to passwall